reactor programslist (also makes it more testable)
