+++ src/containers/journey-page/ProgramsList.tsx
@@ -10,13 +10,11 @@ import progressJourney from "@/actions/progress-journey";
 import Card from "@/components/Card";
 import { Prisma, Program, SwimExercise } from "@prisma/client";
 import InteractiveCardAnimation from "@/animations/InteractiveCardAnimation";
-import Modal from "@/components/Modal";
-import ExerciseSection from "../plans-page/ExerciseSection";
-import Button from "@/components/Button";
 import ApplauseButton from "@/components/ApplauseButton";
 import { isNextDay } from "@/utils/dayjs";
+import HabitTrackingModal from "./HabitTrackingModal";
 
-type ProgramPayload = Prisma.JourneyGetPayload<{
+export type ProgramPayload = Prisma.JourneyGetPayload<{
   include: {
     swimCategory: {
       include: { programs: { include: { swimExercise: true } } };
@@ -85,22 +83,6 @@ export default function ProgramsList({
     [completedProgramIds, currActiveProgramId],
   );
 
-  const exerciseSections = useMemo(() => {
-    const sections: JSX.Element[] = [];
-    // es2015
-    // for (const [type, exercises] of exerciseMap) {
-    //   sections.push(
-    //     <ExerciseSection key={type} type={type} exercises={exercises} />,
-    //   );
-    // }
-    exerciseMap.forEach((exercises, type) => {
-      sections.push(
-        <ExerciseSection key={type} type={type} exercises={exercises} />,
-      );
-    });
-    return sections;
-  }, [exerciseMap]);
-
   const handleClickOpenProgram = (program: Program) => {
     if (
       // disable if the clicked program is not the current active program
@@ -161,6 +143,11 @@ export default function ProgramsList({
     progressJourney();
   };
 
+  const handleCloseModal = () => {
+    setIsOpen(false);
+    setCurrSelectedId(null);
+  };
+
   useEffect(() => {
     if (currActiveProgramRep === 0) {
       setCurrentRep(0);
@@ -178,84 +165,20 @@ export default function ProgramsList({
           <ApplauseButton />
         </Card>
       )}
-
-      <Modal
+      <HabitTrackingModal
         isOpen={isOpen}
-        onClose={() => {
-          setIsOpen(false);
-          setCurrSelectedId(null);
-        }}
-      >
-        <Card className="flex h-full flex-col gap-4">
-          <div className="h-full overflow-y-scroll">
-            <div className="flex flex-col gap-2">{exerciseSections}</div>
-            <h3 className="text-lg font-semibold">
-              Total: {totalDistance}
-              {unit}
-            </h3>
-          </div>
-          <Card className="flex flex-col items-center justify-center gap-2">
-            <p>
-              Do it {selectedProgram && selectedProgram.reps} times to complete
-            </p>
-            <div className="flex items-center justify-center gap-2">
-              {selectedProgram &&
-                new Array(selectedProgram.reps).fill(0).map((_, i) => {
-                  const isRepCompleted =
-                    // show completed icon if the current iteration is less than
-                    // the user's current program rep and the selected program
-                    // is the same as the active program
-                    (i + 1 <= currentRep &&
-                      currActiveProgramId === selectedProgram.id) ||
-                    // or is completed program
-                    isSelectedACompletedProgram;
-
-                  return (
-                    <div
-                      key={i * Math.random()}
-                      className={`flex h-8 w-8 items-center justify-center rounded-full p-1 ${isRepCompleted ? "bg-gray-50" : "bg-gray-400"}`}
-                    >
-                      <p className="text-center">
-                        {isRepCompleted ? "✔️" : i + 1}
-                      </p>
-                    </div>
-                  );
-                })}
-            </div>
-          </Card>
-          <Button
-            // when modal is open
-            isDisabled={
-              // disable if the selected program is the same as the active program AND the current program reps done is the same as the goal/total program reps
-              (currSelectedId === currActiveProgramId &&
-                selectedProgram &&
-                selectedProgram.reps === currentRep) ||
-              // or disable if the selected program is a completed program (included in the completed programs array from the DB)
-              isSelectedACompletedProgram ||
-              // or disable if the current active program is not the same as the
-              // selected program (since we don't want the user to trigger
-              // future progress from previous states e.g. (week 1 button could
-              // progress week 3))
-              currActiveProgramId !== (selectedProgram && selectedProgram.id) ||
-              !isNextDay(lastCompleted)
-            }
-            className="h-16 w-full"
-            onClick={handleClickProgressJourney}
-          >
-            {!isNextDay(lastCompleted) && "Today's goal completed :)"}
-            {isNextDay(lastCompleted) && // show completed if the selected program is the active program and the completed reps is the same as the goal reps
-              ((currSelectedId === currActiveProgramId &&
-                selectedProgram &&
-                selectedProgram.reps === currentRep) ||
-              // or show completed if the active program is not the same as the selected program and the selected program is a completed program
-              (currActiveProgramId !==
-                (selectedProgram && selectedProgram.id) &&
-                isSelectedACompletedProgram)
-                ? "Completed"
-                : "I have done this today!")}
-          </Button>
-        </Card>
-      </Modal>
+        onClose={handleCloseModal}
+        exerciseMap={exerciseMap}
+        totalDistance={totalDistance}
+        unit={unit}
+        selectedProgram={selectedProgram}
+        currentRep={currentRep}
+        currActiveProgramId={currActiveProgramId}
+        isSelectedACompletedProgram={isSelectedACompletedProgram}
+        currSelectedId={currSelectedId}
+        lastCompleted={lastCompleted}
+        onClickJourneyButton={handleClickProgressJourney}
+      />
     </Card>
   );
 }
